pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "langchain_customer_service_robot"
version = "1.0.0"
description = "基于 LangChain 构建的智能客服机器人系统"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-text-splitters>=0.3.0",
    "langchain-deepseek>=0.1.0",
    "langgraph>=0.2.0",
    "faiss-cpu>=1.7.4",
    "rank-bm25>=0.2.2",
    "numpy>=1.24.0",
    "openai>=1.0.0",
    "tiktoken>=0.5.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "websockets>=11.0",
    "jinja2>=3.1.0",
    "pypdf>=3.0.0",
    "python-docx>=0.8.11",
]

[tool.setuptools]
packages = ["langchain_customer_service_robot"]
package-dir = {"langchain_customer_service_robot" = "."}
